Refactor: move tokenInfos from constant to config
