+++ config/tokenInfos.js
