+++ routes/controller/index.js
@@ -11,7 +11,7 @@ const Grid = require("../../models/CustomGrid");
 const getPastISOTime = require("../utils/getPastISOTime");
 const ACCESS_LEVELS = require("../../constants/accessLevels");
 const { OK, UNAUTHORIZED } = require("../../constants/statusCodes");
-const { generateToken, verifyToken } = require("../utils/tokens");
+const { generateToken, verifyToken } = require("../helpers/tokens");
 
 async function getCondition(req, res, next) {
   try {
